App: only set defaultValue in schema editor
The editor matains its own state and handles auto-formatting better if we let it
be on its after the initial load. The onChange callback still sets our state
schema totally fine when changes are valid.

Later on this will require unmnounting the editor and doing some forceful
changes when we want to load a new schema file.

Also the auto format button is probabaly not needed now